+++ lib/generators/kubik/media_library/templates/app/admin/kubik_media_uploads.rb
@@ -1,0 +1,9 @@
+# Host app can take full control of ActiveAdmin registration when auto_register is disabled.
+#
+# KubikMediaLibrary.configure do |config|
+#   config.auto_register_active_admin = false
+# end
+#
+# KubikMediaLibrary::ActiveAdmin::Registration.register_media_upload! do
+#   menu parent: 'Site', priority: 3, label: 'Assets'
+# end